Establish wireless serial communication with Arduino using an ESP8266
Our Johnny-Five host can now wirelessly connect to the ESP8266 (ESP-01)
client. Using wireless serial communication, the Johnny Five library can
dispatch Firmata commands to the ESP8266, which can forward transmissions
to the connected Arduino Uno via the TX and RX pins respectively.

The ESP8266 is capable of being flashed with StandardFirmataWifi and
running as a standalone device. That being said, the tethered Arduino
Uno is a temporary solution to a painful limitation with the chip - it
only has 4 GPIO pins! This might be fine in the end, since the current
stepper driver only requires 4 GPIO pins; but it'll be nice to swap this
tethered Arduino Uno + ESP-01 combo out for a single cheaper, simpler
and more development-friendly NodeMCU board.