Simplify hardware by replacing Arduino / ESP8266 with ESP-12F
Our client code now runs directly on an ESP-12F rather than the Arduino
Uno and ESP8266 connected wireless serial device. This simplifies the
hardware design, enables easier developer and programming, and lowers
the power usage and cost.

The Johnny Five library can dispatch Firmata commands to the ESP-12F,
which can handle the command directly using its many GPIO pins.

The ESP-12F is capable of being flashed with StandardFirmataWifi and
running as a standalone server. Yes, sever...  not a client. That
terminology misunderstanding burned a lot of debugging time. The Firmata
firmware will be committed with the required configuration.