Test integration of 28BYJ-48 stepper motor with a ULN2003 driver
Starting small, the 28BYJ-48 stepper motor provides a level trade-off
between cost, quality and torque. These changes introduce an Arduino
sketch which allows for driving the stepper using a ULN2003 four-phase
hardware stepper driver.

The ULN2003 driver consists of an array of seven open collector
Darlington transistors. Voltage is supplied in series to each coil of
the stepper. This chip is neat because it allows for the low-voltage
microcontroller to switch the 5-12v DC connection, but with each
transistor connected together to exploit the amplification effect of
transistors.

In a real-world test, this motor was capable of pulling the blinds down.

Unfortunately the pulley broke on the way down, and I was unable to test
the heavier "up" direction. How fun! More testing to follow...